AicesPlusOneAgent/requirements.txt
# Google ADK / Genkit
google-genai
google-generativeai>=0.4.0,<0.5.0
google-api-core<2.11.0
google-auth>=2.15.0

# Pydantic for data validation
pydantic>=2.0.0

# HTTP Client
httpx
aiohttp

# JSON Schema Validation
jsonschema

# Environment Variables
python-dotenv

# Utilities
typing-extensions

# Testing
pytest
pytest-asyncio
